requirements.txt
ultralytics
torch
opencv-python-headless
pillow
google-generativeai
boto3
numpy
